+++ src/main/java/com/georgesykes/tilltechtest/ItemFactory.java
@@ -2,7 +2,7 @@ package com.georgesykes.tilltechtest;
 
 public class ItemFactory {
 
-  public Item getItem(String name, Float price) {
+  public Item getItem(String name, Double price) {
     return new Item(name, price);
   }
